maskpass: mask only when there's a password
Instead of masking all the URLs being passed to the Mask function, be
conditional on the masking.

If it does inconditional masking then if there's no password (even no
user), then it will add something that was not there in the first place
and this may create issues on debugging, which is mainly the reason why
you will want to use something like maskpass in the first place.

Additionally, the documentation is fixed. Incorrect explanations like
"use the method StringNE" when the correct method is called "String".